+++ app/src/main/java/fr/wiiznokes/horloge11/utils/Alarm.java
@@ -136,11 +136,11 @@ public class Alarm implements Serializable{
     }
 
 
-    private int id;
-    public int getId() {
+    private long id;
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
